Bump version to 2.6.4 test

+++ sequences/scripts/generate_3d_radial_me_bssfp.py
@@ -30,7 +30,7 @@ def main(
     rf_duration_s: float = 0.5e-3,
     rf_time_bandwidth_product: float = 4.0,
     rf_apodization: float = 0.5,
-    rf_slr_sharpness: float = 1.0,
+    rf_slr_sharpness: int = 1,
     rf_custom_waveform_hz=None,
     rf_custom_raster_s: float | None = None,
     rf_custom_flip_angle_deg: float | None = None,
