+++ sequences/scripts/generate_ute.py
@@ -24,7 +24,7 @@ def main(
     rf_duration: float = 1e-3,
     rf_time_bandwidth_product: float = 2.0,
     rf_apodization: float = 0.5,
-    rf_slr_sharpness: float = 1.0,
+    rf_slr_sharpness: int = 1,
     rf_custom_waveform_hz=None,
     rf_custom_raster_s: float | None = None,
     rf_custom_flip_angle_deg: float | None = None,
